fix adf-delete directive NodeResult

+++ lib/core/directives/node-delete.directive.ts
@@ -110,7 +110,7 @@ export class NodeDeleteDirective implements OnChanges {
 
         let promise;
 
-        if (node.entry.hasOwnProperty('archivedAt')) {
+        if (node.entry.hasOwnProperty('archivedAt') && node.entry['archivedAt']) {
             promise = this.alfrescoApiService.nodesApi.purgeDeletedNode(id);
         } else {
             promise = this.alfrescoApiService.nodesApi.deleteNode(id, { permanent: this.permanent });
